// packages/plugin-assets-ui/src/common/types.ts
export type IAssetCategoryTypes = {
  _id: string;
  name: string;
  order: string;
  code: string;
  parentId: string;
  description: string;
  status: string;
  attachment: IAttachment;
  isRoot: boolean;
  assetCount: number;
};
// ...
export interface IAssetCategory {
  _id: string;
  name: string;
  order: string;
  code: string;
  description?: string;
  attachment?: any;
  status: string;
  parentId?: string;
  createdAt: Date;
  assetCount: number;
  isRoot: boolean;
}

export interface IAsset {
  _id: string;
  name: string;
  type: string;
  categoryId: string;
  parentId: string;
  description: string;
  order: string;
  code: string;
  unitPrice: number;
  customFieldsData?: any;
  createdAt: Date;
  vendorId?: string;

  attachment?: any;
  attachmentMore?: any[];
  assetCount: number;
  minimiumCount: number;
  category: IAssetCategory;
  parent: IAsset;
  childAssetCount: number;
  vendor?: ICompany;
  kbArticleIds: string[];
  knowledgeData?: any;
}